+++ frontend/src/routes/subRoutes/adminRoutes.tsx
@@ -6,22 +6,15 @@ import { PageLink } from "../../components/PageLink";
 import { Users } from "../../pages/Admin/Users";
 import { Contacts } from "../../pages/Admin/Contacts";
 import { Subcontractors } from "../../pages/Admin/Subcontractors";
+import { Suppliers } from "../../pages/Admin/Suppliers";
+
 /**
  * Routes for Admin pages.
  */
 const routes = [
   <Route key="admin" path="/admin" element={<ProtectedRoute component={Admin} />}>
     <Route path="contacts" element={<ProtectedRoute component={Contacts} />} />
-    <Route
-      path="suppliers"
-      element={
-        <ProtectedRoute
-          component={() => {
-            return PageLink("Suppliers");
-          }}
-        />
-      }
-    />
+    <Route path="suppliers" element={<ProtectedRoute component={Suppliers} />} />
     <Route path="subcontractors" element={<ProtectedRoute component={Subcontractors} />} />
     <Route
       path="resources"