+++ frontend/src/types/table.types.ts
@@ -1,0 +1,10 @@
+export interface column {
+  id: number;
+  field: string;
+  headerName: string;
+  width: number;
+}
+
+export interface tableData {
+  data: Array<Object>;
+}